requirements.txt
torch
numpy
scikit-learn
pandas
matplotlib
seaborn
dvc
dvc-gdrive
pyedflib
opencv-python
jupyter
ipywidgets 
widgetsnbextension 
pandas-profiling
sleepecg
wandb
hydra-core
mne
python-dotenv
torchinfo